// src/types/EelvlBlock.ts
export type EelvlBlock = {
  blockId: number
  intParameter?: number
  portalId?: number
  portalTarget?: number
  signText?: string
  signType?: number
  worldPortalTargetWorldId?: string
  worldPortalTargetSpawnPointId?: number
  labelText?: string
  labelTextColor?: string
  labelWrapLength?: number
  npcName?: string
  npcMessage1?: string
  npcMessage2?: string
  npcMessage3?: string
}